Reorganize tests folder
The tests folder was starting to become difficult to manage, too many
scripts many of which would not even work if run in the wrong context.

This split into container and non-container scripts should at least
prevent the user from running scripts in a context where they can't
possibly be expected to work.