Use git archive for docker image
We use a transparent mount in order to provide the container
live access to the working directory. This means that it doesn't matter
so much what's in the docker image for the development flow.

This change dramaitcally reduces the total build time by excluding things
like dependency caches, node_modules, and build directories from being
copied over into the docker container state.

This does affect the all up testing flow, requiring that changes be
checked in before they can be tested. But it's still faster to check
in your changes than it is to wait 45 seconds for docker to copy in
all the useless build context.