Allow additional submissions on events that have already occured
This code was preventing valset updates from working. Namely because
if a valset update was 'Observed' no new confirms could be submitted.

Now in theory the relayer could have submitted that valset confirm but
at the moment it's designed to actually have all validators sign off.

Conceptually either implementation is correct but in this case I think
we want everyone to be able to continue to submit confirms after an event
is confirmed for slashing conditions and other rewards.