Break out ethereum ABI encoding into it's own file
This patch moves the complicated process of encoding validator
set update and batch signatures into the message_signatures module
which allows for easy integration testing of these functions.

Also added in this commit is one integration test for batches and
valsets each comparing the encoded value to a known hash produced
in the Solidity contract

The final change in this patch is to fix batch signature encoding
previously batches where encoded with arrays of byte32 values which
is incorrect.