Attempt signed batch submission
This commit fixes the MsgConfirmBatch message definition on the SDK side and adds
tests for batch submission up to the point where we submit the signed batch
to the cosmos module.

Due to known issues with the Cosmos SDK modules handling of signature submissions
this is as far as we can go right now.