Combine ValsetSignature storage with MsgValsetConfirm storage
Orignially there was a flow for valset confirmation that involved
storing the entire MsgValsetConfirm, Alex then started to remove this
flow and instead store only the signature itself in an index that was
designed to easily prefix search.

The previous several commits focused on removing the new generic flow
for signature submissions. This patch finishes that work by combining
the two flows and adding a few new tests to make sure the endpoints
work.