Redefine BatchConfirmMsg with two Eth addresses
The the batch confirm needs two eth addresses, one that contains the
signer ethereum address for use by the relayer to determine how to
order signatures for submission to the Peggy contract and another to
indicate the token contract we are trying to confirm with this signature
submission.

Previously the go code was expected the token contract address while
the rust code was sending the signers address.