More reliable legacy transactions
It seems that the error message when a transaction fails to get into
the blockchain has failed and is now simply a server error 500 response
code.

I'm not 100% sure that thi is the case, but I also don't care to dig too
deeply into the leagcy system. So I've just added this as a retry condition
is contact and updated the dependency here.