Update PeggyID
Right now the cosmos chain is using a value of defaultpeggyid as it's
peggyid. Since this needs to be synced end to end we have to change
the default value in the tests for now.

Eventually we want to correct this so that it is retrieved from the
cosmos chain by the deployment script, at that point everything
will use the cosmos chain as the source of truth and we'll be
golden.