Add test for msgSendToEth
In an effort to better understand my problems with this message I
designed a test that doesn't seem to find anything wrong with it.

This leads me to believe that the error I'm having trouble with is
somewhere in the antehandler before we get the message in our code
or in my own message sending code.