Remove the last of bridge signature submission
Alex made an effort to generalize bridge signature submissions into
a generic 'bridge signature submission' message. This is mostly just
extra interfaces and complexity because Peggy only needs two different
types of signatures. It's possible adding an arbitrary transaction
endpoint would add a third type of signature, but we're never going
to have dozens of these.

The extra handling logic involved in generalizing the process is simply
not justified.