Better Orchestrator error messages
This patch to Contact displays the raw_log error message from cosmos
as part of the user error message. This is a big improvement in
debugging ergonomics, the previous way of getting this info required
using the trace logging level and carefully specifying the correct
module as a filter statement to the logging library.