Install vim-minimal in Docker image
A conflict between vim-minimal package and vim-enhanced (which seems to be
the default alias for 'vim' seems to have started after the update to
the Fedora 33 container image, which now contains the minimal vim
rather than simply 'vi'.

This simple patch fixes the issue by specifying only a dependency
on vim-minimal. You could also just remove the requirement assuming
some future version of the image won't revert to vi. Which I don't think
is probable.