Get PeggyID from blockchain params on deploy
This patch modifies the contract deployer script to retrieve the
PeggyID from the chain parameters using a new ABCI endpoint also
included in this patch.

This allows for the chain to select a PeggyID either by placing it
into a genesis file or by voting on it as a governance paramater
after the chain starts but before the Peggy contract is deployed.

On deployment the contract deployer will retrieve the PeggyID from
the chain and the value will then be fixed in the contract storage.

At some point we should probably forbid deployments of the defualt
peggy ID in the deployment script outside of test mode. To prevent
possible failures where validators on multiple peggy chains find
signatures intended for one Peggy contract being used against another.