Implement relayer gas optimizations for valsets
The valset relaying routine will update the valset if there is a new one
regardless of cost or how many new ones there are. This obviously isn't
sustainble on production Ethereum becuase it represents a wallet
draining bug. Anyone can make an unpermissioned validator set request
as many as they like, potentially bankrupting the relayer by having
them update the validator set at a cost of $4 in gas for every cosmos
block.

This patch uses the percentage change checking logic to compute if
more than 1% of the power has been reallocated, if it's less than that
no validator set is submitted.

This does present problems for the tests, which expect validator set
updates happen but don't do anything to change the power of the staking
validators. Until this can be remedied the tests will be run with this
optimization disabled.