Eth Oracle picks up at the correct starting block
This patch makes several changes to allow the Ethereum Oracle to
correctly maintain event nonce order as it plays events from Ethereum
into the Peggy module.

1) Add gRPC endpoint to retrieve the last event nonce relayed to the
   Peggy module by a spefific validator.
2) Update web30 to better handle the task of searching the entire
   Ethereum blockchain history for an event matching that last event
   nonce, allowing the oracle to find the correct block to continue
   from
3) Modified the Oracle to recover from crashes and other wonky states
   where events in a block are partially relayed by filtering out events
   that are greater than the last seen event nonce.

Flaws in this implementation mainly revolve around the inefficiency of
finding the last checked block. It grabs every event in the history
of the contract and iterates until it finds the correct one. This won't
work once the contract has a long enough history that the response
becomes very large (5 megabytes is the hardcoded value now but whatever
value is set the number of events will eventually exceed it, not to
mention the extra load on the Ethereum node to serve it).

This will have to be corrected with some form of binary search
eventually.