Order sigs using the old valset not the new valset
This is a very subdle but critical error. The signatures need to
be ordered based on the current members of the validator set not
the upcoming members. Otherwise any new member in the new validator
set will be submitted as an argument to be compared to the old
validator set. The contract will then reject such a transaction.