Outgoing tx batch slashing
this is very similar to validator set slashing except that we need
to add a block height paramater. Validator set updates have their
nonces created based on the present block height, whereas batches
are simply sequential.

This should probably be changed for valset updates and a block
height added there.

Co-authored-by: Justin Kilpatrick <justin@althea.net>