Use ValsetUpdateEvent for submissions
The Peggy contract works by accepting a list of current validators and
powers that is then compared to a hash in storage. This removes the need
to actually store the whole valdiator set on Ethereum and instead it is
stored in the much cheaper Ethereum event feed.

Previously the relayer relied heavily on the assumption that relayers
would only submit the exact validator set stored on Cosmos. This is not
required to be the case. Specifically around sorting it's possible for a
relayer to submit a validator set in arbitrary sorting order, resulting
in a stored hash on the Ethereum blockchain depending on that arbitrary
sorting.

Another edge case is the deployment of the Peggy contract, for which no
validator set is stored on Cosmos at all. The weak assumption that
validator set 0 was the same as the current validator set broke Peggy
testnet #1 for several hours.

The final case where the valset maybe wrong is in the case of bridge
highjacking or just plain old incorrectness on the Cosmos chain.
Returning the wrong validator set or having one created outside of the
normal channels and submitted are all possibilities.

This patch implements the correct solution to this whole class of
problem, just grab the exact validator set already submitted from the
event history and use that. Bad soring is fine, no valset on Cosmos is
fine, and even an attempted bridge takeover may be automatically
smoothed away if the hostile validators left their orchestrators
running.