Add height to valsets
This removes our dependency on the validator set nonce as a proxy for
the block height at which it was issued.

Without this Peggy would need to have strange requirements on chain
restarts to ensure that historic validator sets where not interpreted as
happening far in the future and other such bad news.

Eventually we should move to simply using an incrementing counter for
validator set nonce rather than the block height. Although with this out
of the way there's no particularly pressing need to do so.

Co-authored-by: Justin Kilpatrick <justin@althea.net>