Security checks for claims and valset requests
This adds checks to permission MsgWithdrawClaim and
MsgDepositClaim. Without these votes will reach the actual
attestation logic and maybe cause it to panic if the submitter is not
actually a validator.

Also permissioned in this commit is MsgValsetRequest which is not a
strong security requirement but has few downsides in exchange for some
upsides against spam attacks.

Co-authored-by: Justin Kilpatrick <justin@althea.net>