Remove misguided total power check
This check was intended to warn the user when there where non-normalized
validator sets being submitted. But as a way to check for that this code
doesn't work. Of course the power difference can be greater than
u32::MAX! Imagine if the validators sets where 100% different the value
would be 2 * u32::MAX since there is absolutely no shared power!

This isn't really the right place to check that validator sets don't sum
to the wrong amount of power. The real job of maintaining these sorts of
invariants is on the Cosmos blockchain.