Upgrade docker test env with go 1.15
Fedora has not yet packaged go 1.15 so now we have to pull it in
manually using tar steps and then set some env vars ourselves.

This is required because tendermint as of the Stargate RC5 upgrade now
depends on the newer version of go. I don't think following the latest
version of a language so closely is generally good practice but that's
their choice.