Modfiy validator powers in tests
this patch changes the integration tests to actually modify the
validator powers in order to generate new validator sets. Previously we
could simply send in a ValsetUpdateMsg and trigger the creation of an
update to relay. But now that this is done automatically by the Peggy
module we have to change the power intentionally to get the same effect.