Rename validator to orchestrator in peggy messages
It was decided to use Orchestrator universally for all non-relayer
tasks. This is somewhat controversial as the 'orchestrator' is just a
name we use to refer to the two roles a validator running Peggy must
participate in. The signing of ethereum messages (for things going to
Ethereum) and the submission of oracle messages (for things coming from
Ethereum).

This combined single name is useful but may be confusing to the
uninitiated.