Fix validator out test
In order to generate validator sets for validator set update tests
we vary the validator set power randomly. For the validator out test
this isn't feasible because we use only 3 validators in our test
environment. Meaning even a 1% allocation to the down validator will
result in the remaining online validators not having enough power to
operate the bridge.

This is a good argument for using a 4 or 5 validator set in CI but we
are already resource constrained enough. So for now lets just disable
the validator set update testing part of the validator out flow and
replace it with a simple check that the validator set on Ethereum is
nonzero.