Add batchTimeout to Peggy.sol
This adds a batch timeout to the Peggy.sol contract, every batch is
produced with a timeout height some number of blocks into the future.

Once that block height is reached the batch is no longer valid and can
not be submitted.

Test cases are of course updated to include the new throw condition.