Fix consensus issue with attestation processing
You can't iterate over maps in CosmosSDK becuase of the ranomness in the
HashMap algorithm. So what we do instead is take the keys, sort them and
then perform key lookups for each entry. This isn't that much slower
than before especially since we only do the sort once per block.

The other change made here is to correct a few issues around the 'early'
exit behavior. The previous behavior was both not optimal and incorrect
in a purely functional sense.

After experimenting with several implementations of this optimization I
can't find any that work in all cases. So it will be dropped for now.