Create full batches in tx stress test
This test expands the tx stress test to live up to it's name
BATCH_STRESS and actually send batches back to Ethereum.

There are a number of challenges to this, mostly that Cosmos is very
slow about including transactions in 'block' mode although this may
be the fault of our own agressive retry.

Also included are some minor changes to the send to eth function to
better check the command.