Don't print error on signature status
This function contains all the logic for making sure that signatures are
actually possible to submit and otherwise valid. So it makes sense to
use it to check that signatures can work in the first place.

This error message is then very disruptive to the user becuase it seems
like there's a major failure while things are actually just being
evaluated.