Sign multiple valsets in the same message
This modifies both the Go module endpoint and the Orchestrator code
in order to allow for much faster validator set update signing.

I suspect that validator set updates are being produced dramatically
too often on the Cosmos side currently, but since it's possible to
create a new validator set every block we need to be able to keep
up with that sort of speed.