Update validator set relaying logic
The previous logic for the validator set relaying was not nearly robust
enough to deal with a constant barrage of validator sets.

First and foremost the 'latest' validator set endpoint only returns the
last 5 validator sets. If a new set is generated every block there's no
way that every single validator will manage to submit their signature
within a 5 block window.

Even worse we previously made no attempt at finding a historical
validator set that could move between the current validator set in the
bridge and the present validator set.

This new code is much more computation and request intensive but solves
these problems, we look back through the entire history of valdiator
sets and compare to the present validator set in an attempt to find the
newest one with macthing signatures.