Relayer should reuse the current valset
The last couple of patches concede that we need to know the latest
validator set in the contract on Ethereum before we can make any relayer
decisions. It's not enough to simply assume that the latest set on
Ethereum is close enough to the latest set on the Cosmos chain, we have
to handle all possbilities.

Getting the latest validator set is rather intense, we search the
Ethereum block history 50k blocks at a time until we manage to find what
we're looking for. It's best to re-use that effort by placing the search
at the very top fo the relayer loop. Once we have the data we can
perform all of our batch and validator set relaying without having to
redo the work or handle not finding the current validator set in
multiple places.