Print error in Oracle resync
I believe that a common testnet error came from users not setting their
delegate keys, they would then get this error message without any clear
specification if it was actually the GRPC connection or the lack of a
registered address.

Adding this into the print statement will produce errors that are much
easier for me to debug remotely.