Fix bogus batch relaying error
This was missed in "Relayer should reuse the current valset" what was
previously an error condition became the condition for there *not* being
a batch to relay. Obviously constantly printing an error when there's
nothing wrong and in fact nothing to do at all makes a lot of console
noise.