Fix double jail if multiple slashing conditions are met
**Problem:**
If validator missed attesting both ValsetRequest and BatchRequest (due to downtime) which are created at same height, we would try to slash same validator twice as both slashing conditions are met. To fix this we are no longer using a single copy of the current bonded set for all conditions, but instead grabbing a new copy of the 'currently bonded set' inside of each condition. Meaning a validator jailed by the condition above will no longer be in the bonded set for the next condition.

Note that currently only bonded validators are being slashed and a validator in the unbonding state will not be slashed for failing to sign validator sets. This goes against our implementation requirements of having validators sign for some period (current thinking is 24 hours) after they start unbonding.

**Solution:**
- [x] Slash only once if multiple slashing conditions(Valset, Batch) are met at same block height.
- [x] Precedence Order of slashing - ValsetSlashing > BatchSlashing > ClaimSlashing