Improve error message for incorrect delegate keys
The delegate keys warnings failed to take into account that an error
response is the correct behavior if the key has not been found. So the
displayed help was not the best.

This patch improves the displayed help to deal with this situation in
additon to the various possiblities of confused delegate keys