Default to log level 'info'
In general panics for simple validation errors result in a lot of
useless data being dumped to the screen when we just want to give the
user a one line message.

But if the user hasn't enabled logging they may just see their process
exit with error code 1, silently failing for no reason they can see.

In order to resolve this we should ensure that the default logging level
is info, so that users who fail to enable logging themselves still get
somthing.