Distinguish between syncing and waiting to start
Previously I thought that 'syncing' remained true while a node was
waiting for the chain to start, sadly that's not the case instead we
have the much stranger condition of 'syncing' being true only until the
first prevote comes in. How this is determined I don't even want to
think about too much. So instead we now check the latest block endpoint
which returns null until there is in fact a block to look at.