Update client for GravityV2 features
This is a first draft of support for GravityV2 in our testing client.
The biggest thing that makes this a 'first draft' is the lack of proper
metadata support. I need to determine how to query decimal metadata and
more reliably determine if an asset originated on Cosmos or Ethereum.
That's going to require some endpoint changes I think.