Remove claims slashing
After much discussion there seems to be a soft agreement that Claims
slashing is at best useless and at worst a threat to the Cosmos chain
running it.

If a group of validators can make false claims, they can also take over
the chain, as they must have 66% of the total voting power. So slashing
them won't exactly stop them from doing anything.

If a validator make an incorrect claim, due to an ehtereum fork or any
other unintentional reason they will be slashed. Potentially causing
chaos and even taking down the chain running Gravity if there is some
unexpected fork or other bad behavior on the EVM chain.

So we have an attack that slashing doesn't prevent and a worst case that
could halt the chain through no fault of any validator. When put in
those terms this condition simply doesn't make sense and should be
removed.