Remove OutgoingTx type
Previously we had two different types managing transactions in the pool
OutgoingTx and OutgoingTransferTx the former was used for state sorage
in the tx pool and the latter was used for tx's in side of actual
batches.

This patch removes OutgoingTx and replaces it with OutgoingTransferTx
which contains more data.

Another choice made here is to only represent bridge transactions in the
store using the ERC20 token type, rather than PeggyDenoms. It makes
sense to attach the ERC20 address even for a Cosmos originated address
and generally seems to keep useless conversions down.