Rename BatchFees sum field
This patch renames the field for the sum of all transaction fees in a
batch for the BatchFees message from topOneHundred to totalFees. The
fact that there is a one hundred tx limit on batches currently is by no
means a permenant decision and that constant value could change at any
time.

So it's probably better to have a generic field name and change it
sooner rather than later.