Upgrade to Contact 0.4.2 and Web30 0.10.8
This patch updates to the latest contact version. Fixing some block
parsing issues and response size issues we're seeing as the Althea
testnet continues to scale up.

Web30 was having issues with log sizes, this commit also reduces the
search block interval from 50k to 5k combined with the size increase in
response buffers this should be more than sufficient.