Fix Rust and Solidity from Peggy -> Gravity rename
The Rust code had all mentions of Peggy corrected properly. Including
imports, but the files themselves where not renamed, breaking the
imports.

Likewise references to Peggy in the solidity where renamed but the
contract itself was not. Fortunatley these seem to be pretty small
changes to get everything working again.