Expand orchestrator keys test
This test still does not work and I've decided that it needs to be
dropped until after upgrading to proto tx. After determining
conclusively that the issue does not originate in the Go code (see the
previous commit) this seems to be attributible to proto issues or some
other race condition.