Add support for custom prefixes
This patch adds support for custom prefixes. To allow for chains with
different address prefixes than the default 'cosmos'.

Sadly this requires the user to specify the prefix as an argument when
launching the orchestrator. There's no way around this until a future
version of Cosmos exposes the prefix to clients. Looking upstream the
work for this seems to be in progress.