Oracle for valset rewards
This patch adds oracle support for validator set update rewards.

What this does functionally is allow us to keep the number of tokens
in the keeper consistent with what we actually have in the bridge. Even
if we're issuing tokens on Ethereum for the first time or spending
Ethereum denominated assets to fund validator set updates.