Failures in OutgoingTxBatchExecuted should panic
A failure in the function OutgoingTxBatchExecuted will cause a double
spend or cause users to lose tokens. Either a later batch cleanup
succeeds and tokens are burned without actually being sent on Ethereum
or a batch timeout occurs for a batch that has already executed and
users are able to withdraw their funds while also having their funds on
Ethereum.

Either case is unacceptible so we should panic.