CancelOutgoingTxBatch should only cancel batches of the same token type
This is a pretty obvious bug that I had to add a test to find. The code
in OutgoingTxBatchExecuted assumed that we had the same token type in
the outgoing tx batch iterator, but that was not the case, therefore it
would try and delete batches in indexes that didn't exist.

This doesn't actually produce a correctness bug in the sense of minting
or burning tokens, but is nice to have out of the way.