Add more logging for signature failures
This patch adds some debug logging for the persistent signature failures
problem that I've been chasing for a while. My theory is that we are
just not encoding specific batches and validator sets correctly. This
issue comes up rarely enough that including logging for it in the main
tree is really the only way we'll catch it.