Pass maxElements through batch fees function tree
Previously we used the constant OutgoingTxBatchSize deep inside this
call stack. Leading some people to believe that the fees function did
not in fact respect batch size.

The only BuildOutgoingTXBatch has a configurable number of elements if
to faciliate the way some tests are designed. Arguably it should use the
constant as well. But instead for this patch we've simply reordered the
entire tree to the value.