More clarity for highjacking detection messages
This patch corrects an error message which implied that a bridge attack
was taking place when in fact we have a logical error where pieces of
info that are supposed to always agree do not.

The new code is modified to not request the latest validator set nonce
from two sources (possibly resulting in a race condition) and to also
panic in this same edgecase since it should not be possible.