Clean up params, better default valset reward
This patch changes the valset reward value from a nullable pointer to a
coin to a coin value. This coin is then set to an invalid value to
indicate the 'no reward' case.

The indirection caused by the pointer was causing issues registering
this variable in ParamsSetPairs() that didn't seem easily to resolve.

This patch also ensures that all parameters are registered, in the same
order in all locations, and removes a few orphan keys.