Remove slashing logic in GetLastObservedEventNonce
This logic to avoid slashing for validators is based on claims slashing
which has been removed intentionally to increase stability of the host
chain.

Give that we no longer have the same concerns, we can now simply return
the latest event nonce minus one so that they can vote on the current
event.