Add event pruning cutoff
The addition of a new endpoint for browsing oracle events made clear
the problem that as events execute they get pruned immediately,
preventing the endpoint from being useful for debugging or for frontend
visualization.

This patch adds a cutoff to event pruning that allows for some recent
events to be kept around.