Add imperfect internal -> proto into implementations
For evidence based slashing we need to
be able to convert our internal types back to proto types. The proto
types contain a 'height' value which isn't used in their respective
signatures but is used to deal internally with slashing or other info.

This imperfect conversion really shouldn't implement .into() but it
seems ok here.