Fix for unbonding hook init
We have a hook that runs when a validator starts to unbond. This hook
generates a new validator set update right away to ensure that we are
always up to date with the latest set.

The Evidence based slashing test triggers this hook and revealed that it
was improperly called in app.go and was never working in the first
place.

This patch corrects the init order and adds a null check to prevent it
from happening in the future.

The other conclusion that can be gleaned from this bug is that this hook
is never triggered by jailing performed in the endblock, otherwise all
existing networks and testnets would just panic the first time a user
was slashed.

This patch also adds a manual invocation of the keeper function where we
jail in the endblocker to counteract this.